phase3-django-api/pyproject.toml
[project]
name = "phase3-django-api"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "coverage>=7.10.4",
    "django>=5.2.5",
    "django-filter>=25.1",
    "djangorestframework>=3.16.1",
    "drf-spectacular>=0.28.0",
    "drf-spectacular-sidecar>=2025.8.1",
    "drf-yasg>=1.21.10",
    "factory-boy>=3.3.3",
    "mysqlclient>=2.2.7",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "pytest-django>=4.11.1",
    "python-decouple>=3.8",
]
